agents/rag_agents/marketing_growth/marketing_agent.py
# import os
# import random
# import time
# import logging
# from google.generativeai import configure, GenerativeModel
# from dotenv import load_dotenv
# from agents.rag_agents.marketing_growth.retrieval.marketing_retriever import MarketingRetriever
# from agents.rag_agents.marketing_growth.models.campaign_model import CampaignModel
# from agents.rag_agents.marketing_growth.data_processing.market_analyzer import MarketAnalyzer
# from agents.rag_agents.marketing_growth.data_processing.campaign_optimizer import CampaignOptimizer

# # Load environment variables
# load_dotenv()
# GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

# # Configure Gemini API
# configure(api_key=GEMINI_API_KEY)

# # Configure Logging
# logging.basicConfig(level=logging.INFO)
# logger = logging.getLogger(__name__)

# class MarketingAgent:
#     def __init__(self, knowledge_base_path, hub=None):
#         """
#         Initialize the MarketingAgent with all necessary components.
#         """
#         self.hub = hub
#         self.retriever = MarketingRetriever(knowledge_base_path)
#         self.campaign_model = CampaignModel()
#         self.market_analyzer = MarketAnalyzer()
#         self.campaign_optimizer = CampaignOptimizer()
#         self.performance_score = 0.0
#         self.update_interval = 10  # Default update interval in seconds

#     def run_marketing_campaign(self, product_info, market_query):
#         """
#         Runs a full marketing campaign lifecycle.
#         """
#         try:
#             logger.info("Running marketing campaign...")

#             trends = self.retriever.retrieve_trends(market_query)
#             if "error" in trends:
#                 if "429" in trends["error"]:
#                     trends = {"trends": "Fallback market trend data due to quota exhaustion"}
#                 else:
#                     raise ValueError(trends["error"])

#             message = self.campaign_model.generate_message(product_info, trends) or "Fallback campaign message"
#             score = self.campaign_model.score_campaign(message) or 1
#             optimized_message = self.campaign_model.optimize_campaign_language(message) or "Fallback optimized message"
#             social_ad = self.campaign_model.generate_social_media_ad(optimized_message, "Instagram") or "Fallback social ad"

#             return {
#                 "market_trends": trends,
#                 "campaign_message": message,
#                 "optimized_message": optimized_message,
#                 "campaign_score": score,
#                 "social_media_ad": social_ad
#             }
#         except Exception as e:
#             logger.error(f"Error running campaign: {str(e)}")
#             return {"error": str(e)}

#     def analyze_market_and_competitors(self, query):
#         """
#         Conducts market trend and competitor analysis.
#         """
#         try:
#             logger.info("Analyzing market and competitors...")
#             market_trends = self.market_analyzer.analyze_trends(query) or {"trends": "Fallback market trends"}
#             competitors = self.retriever.retrieve_competitor_data(query) or {"competitors": "Fallback competitor data"}
#             return {"market_trends": market_trends, "competitor_data": competitors}
#         except Exception as e:
#             logger.error(f"Error analyzing market: {str(e)}")
#             return {"error": str(e)}

#     def optimize_campaign(self, campaign_data):
#         try:
#             logger.info("Optimizing campaign...")
#             performance_metrics = campaign_data.get("performance_metrics", {})
#             engagement_rate = performance_metrics.get("engagement_rate", 0)
#             roi = performance_metrics.get("roi", 0)

#             if engagement_rate <= 0 or roi <= 0:
#                 raise KeyError("Required performance metrics missing: 'engagement_rate' or 'roi'.")

#             improvements = self.campaign_optimizer.suggest_improvements(campaign_data) or "No improvements suggested."
#             budget_adjustment = self.campaign_optimizer.optimize_budget(
#                 campaign_data["current_budget"], performance_metrics
#             ) or campaign_data["current_budget"]

#             return {
#                 "suggestions": improvements,
#                 "new_budget": budget_adjustment
#             }
#         except KeyError as ke:
#             logger.error(f"Key missing during optimization: {str(ke)}")
#             return {"error": str(ke)}
#         except Exception as e:
#             logger.error(f"Error optimizing campaign: {str(e)}")
#             return {"error": str(e)}

#     def fetch_campaign_insights(self, topic):
#         """
#         Fetches marketing insights for a given topic.
#         """
#         try:
#             logger.info(f"Fetching campaign insights for topic: {topic}")
#             insights = self.retriever.fetch_marketing_insights(topic) or {"marketing_insights": "Fallback insights"}
#             return {"marketing_insights": insights}
#         except Exception as e:
#             logger.error(f"Error fetching insights: {str(e)}")
#             return {"error": str(e)}

#     def evaluate_campaign_performance(self, campaign_name):
#         """
#         Evaluates historical performance of campaigns.
#         """
#         try:
#             logger.info(f"Evaluating campaign performance for: {campaign_name}")
#             performance_summary = self.campaign_optimizer.evaluate_campaign_performance(campaign_name) or {"summary": "Fallback performance summary"}
#             return {"performance_summary": performance_summary}
#         except Exception as e:
#             logger.error(f"Error evaluating performance: {str(e)}")
#             return {"error": str(e)}

#     def update_performance_metrics(self):
#         """
#         Simulates periodic updates of performance metrics.
#         """
#         self.performance_score = random.uniform(50, 100)  # Simulate random performance score
#         logger.info(f"Updated performance score: {self.performance_score}")

#     def send_performance_data_if_needed(self):
#         """
#         Sends performance data to StrategyLeadershipAgent via CommunicationHub.
#         """
#         if not self.hub:
#             logger.warning("No CommunicationHub connected.")
#             return

#         message = {
#             "request": "performance_update",
#             "department": "MarketingAgent",
#             "performance_score": self.performance_score,
#             "kpis": {
#                 "status": "Below Threshold" if self.performance_score < 75 else "Above Threshold",
#                 "weak_areas": ["Ad Engagement", "Sales Conversion"] if self.performance_score < 75 else [],
#             }
#         }
#         self.hub.send_message("MarketingAgent", "StrategyLeadershipAgent", message)
#         logger.info(f"Sent performance data: {message}")

#     def receive_message(self, sender, message):
#         """
#         Handles incoming messages from agents.
#         """
#         logger.info(f"[MarketingAgent] Received message from {sender}: {message}")

#         if message.get("action") == "reward":
#             logger.info(f"Reward received: {message['reason']}")
#         elif message.get("action") == "penalty":
#             logger.warning(f"Penalty received: {message['reason']}")
#         elif message.get("action") == "improve":
#             weak_areas = message.get("weak_areas", [])
#             guidance = message.get("guidance", "No guidance provided.")
#             logger.info(f"Improvement Plan received: Areas - {weak_areas}, Guidance - {guidance}")
#         else:
#             logger.warning(f"Unsupported message type received from {sender}.")

import os
import random
import logging
from google.generativeai import configure, GenerativeModel
from dotenv import load_dotenv
from agents.rag_agents.marketing_growth.retrieval.marketing_retriever import MarketingRetriever
from agents.rag_agents.marketing_growth.models.campaign_model import CampaignModel
from agents.rag_agents.marketing_growth.data_processing.market_analyzer import MarketAnalyzer
from agents.rag_agents.marketing_growth.data_processing.campaign_optimizer import CampaignOptimizer

# Load environment variables
load_dotenv()
GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")

# Configure Gemini API
configure(api_key=GEMINI_API_KEY)

# Configure Logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class MarketingAgent:
    def __init__(self, knowledge_base_path, hub=None):
        """
        Initialize the MarketingAgent with all necessary components.
        """
        self.hub = hub
        self.retriever = MarketingRetriever(knowledge_base_path)
        self.campaign_model = CampaignModel()
        self.market_analyzer = MarketAnalyzer()
        self.campaign_optimizer = CampaignOptimizer()
        self.performance_score = 0.0

    def run_marketing_campaign(self, product_info, market_query):
        """
        Runs a full marketing campaign lifecycle.
        """
        try:
            logger.info("Running marketing campaign...")

            trends = self.retriever.retrieve_trends(market_query)
            if not trends or "error" in trends:
                trends = {"trends": "Fallback market trend data"}

            message = self.campaign_model.generate_message(product_info, trends) or "Fallback campaign message"
            score = self.campaign_model.score_campaign(message) or 1
            optimized_message = self.campaign_model.optimize_campaign_language(message) or "Fallback optimized message"
            social_ad = self.campaign_model.generate_social_media_ad(optimized_message, "Instagram") or "Fallback social ad"

            return {
                "market_trends": trends,
                "campaign_message": message,
                "optimized_message": optimized_message,
                "campaign_score": score,
                "social_media_ad": social_ad
            }
        except Exception as e:
            logger.error(f"Error running campaign: {str(e)}")
            return {"error": str(e)}

    def analyze_market_and_competitors(self, query):
        """
        Conducts market trend and competitor analysis.
        """
        try:
            logger.info("Analyzing market and competitors...")
            market_trends = self.market_analyzer.analyze_trends(query) or {"trends": "Fallback market trends"}
            competitors = self.retriever.retrieve_competitor_data(query) or {"competitors": "Fallback competitor data"}
            return {"market_trends": market_trends, "competitor_data": competitors}
        except Exception as e:
            logger.error(f"Error analyzing market: {str(e)}")
            return {"error": str(e)}

    def optimize_campaign(self, campaign_data):
        try:
            logger.info("Optimizing campaign...")
            performance_metrics = campaign_data.get("performance_metrics", {})
            engagement_rate = performance_metrics.get("engagement_rate", 0)
            roi = performance_metrics.get("roi", 0)

            if engagement_rate <= 0 or roi <= 0:
                raise KeyError("Required performance metrics missing: 'engagement_rate' or 'roi'.")

            improvements = self.campaign_optimizer.suggest_improvements(campaign_data) or "No improvements suggested."
            budget_adjustment = self.campaign_optimizer.optimize_budget(
                campaign_data["current_budget"], performance_metrics
            ) or campaign_data["current_budget"]

            return {
                "suggestions": improvements,
                "new_budget": budget_adjustment
            }
        except KeyError as ke:
            logger.error(f"Key missing during optimization: {str(ke)}")
            return {"error": str(ke)}
        except Exception as e:
            logger.error(f"Error optimizing campaign: {str(e)}")
            return {"error": str(e)}

    def fetch_campaign_insights(self, topic):
        """
        Fetches marketing insights for a given topic.
        """
        try:
            logger.info(f"Fetching campaign insights for topic: {topic}")
            insights = self.retriever.fetch_marketing_insights(topic) or {"marketing_insights": "Fallback insights"}
            return {"marketing_insights": insights}
        except Exception as e:
            logger.error(f"Error fetching insights: {str(e)}")
            return {"error": str(e)}

    def evaluate_campaign_performance(self, campaign_name):
        """
        Evaluates historical performance of campaigns.
        """
        try:
            logger.info(f"Evaluating campaign performance for: {campaign_name}")
            performance_summary = self.campaign_optimizer.evaluate_campaign_performance(campaign_name) or {"summary": "Fallback performance summary"}
            return {"performance_summary": performance_summary}
        except Exception as e:
            logger.error(f"Error evaluating performance: {str(e)}")
            return {"error": str(e)}
